+++ proto/src/lib.rs
@@ -19,7 +19,6 @@
 
 pub mod backup;
 pub mod cli;
-pub mod config;
 pub mod constants;
 pub mod internal;
 pub mod messages;
