+++ server/core/src/config.rs
@@ -6,7 +6,6 @@
 
 use cidr::IpCidr;
 use kanidm_proto::backup::BackupCompression;
-use kanidm_proto::config::ServerRole;
 use kanidm_proto::constants::DEFAULT_SERVER_ADDRESS;
 use kanidm_proto::internal::FsType;
 use serde::Deserialize;
@@ -245,6 +244,37 @@ impl Display for HttpAddressInfo {
                 }
                 f.write_str("]")
             }
+        }
+    }
+}
+
+#[derive(Debug, Deserialize, Clone, Copy, Default, Eq, PartialEq)]
+pub enum ServerRole {
+    #[default]
+    WriteReplica,
+    WriteReplicaNoUI,
+    ReadOnlyReplica,
+}
+
+impl Display for ServerRole {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ServerRole::WriteReplica => f.write_str("write replica"),
+            ServerRole::WriteReplicaNoUI => f.write_str("write replica (no ui)"),
+            ServerRole::ReadOnlyReplica => f.write_str("read only replica"),
+        }
+    }
+}
+
+impl FromStr for ServerRole {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "write_replica" => Ok(ServerRole::WriteReplica),
+            "write_replica_no_ui" => Ok(ServerRole::WriteReplicaNoUI),
+            "read_only_replica" => Ok(ServerRole::ReadOnlyReplica),
+            _ => Err("Must be one of write_replica, write_replica_no_ui, read_only_replica"),
         }
     }
 }
@@ -509,7 +539,7 @@ impl Configuration {
             origin: Url::from_str("https://idm.example.com")
                 .expect("Failed to parse built-in string as URL"),
             log_level: LogLevel::default(),
-            role: ServerRole::WriteReplica,
+            role: ServerRole::WriteReplicaNoUI,
             repl_config: None,
             integration_repl_config: None,
             otel_grpc_url: None,
@@ -645,10 +675,6 @@ impl ConfigurationBuilder {
 
         if let Some(origin) = &cli_config.origin {
             self.origin = Some(origin.clone());
-        }
-
-        if let Some(role) = &cli_config.role {
-            self.role = Some(*role);
         }
 
         // networking
