+++ server/core/src/https/mod.rs
@@ -1,7 +1,7 @@
 use self::extractors::ClientConnInfo;
 use self::javascript::*;
 use crate::actors::{QueryServerReadV1, QueryServerWriteV1};
-use crate::config::{AddressSet, Configuration, TcpAddressInfo};
+use crate::config::{AddressSet, Configuration, ServerRole, TcpAddressInfo};
 use crate::tcp::process_client_addr;
 use crate::CoreAction;
 use axum::{
@@ -22,7 +22,7 @@ use crypto_glue::{
 use futures::pin_mut;
 use hyper::body::Incoming;
 use hyper_util::rt::{TokioExecutor, TokioIo, TokioTimer};
-use kanidm_proto::{config::ServerRole, constants::KSESSIONID, internal::COOKIE_AUTH_SESSION_ID};
+use kanidm_proto::{constants::KSESSIONID, internal::COOKIE_AUTH_SESSION_ID};
 use kanidmd_lib::{idm::authentication::ClientCertInfo, status::StatusActor};
 use serde::de::DeserializeOwned;
 use sketching::*;
@@ -202,23 +202,31 @@ pub async fn create_https_server(
     maybe_tls_acceptor: Option<TlsAcceptor>,
     tls_acceptor_reload_tx: &broadcast::Sender<TlsAcceptor>,
 ) -> Result<Vec<task::JoinHandle<()>>, ()> {
-    let all_js_files = get_js_files(config.role)?;
-    // set up the CSP headers
-    // script-src 'self'
-    //      'sha384-Zao7ExRXVZOJobzS/uMp0P1jtJz3TTqJU4nYXkdmsjpiVD+/wcwCyX7FGqRIqvIz'
-    //      'sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM';
+    let js_checksums = match config.role {
+        ServerRole::WriteReplicaNoUI => String::new(),
+        ServerRole::WriteReplica | ServerRole::ReadOnlyReplica => {
+            let all_js_files = get_js_files(config.role)?;
+            // set up the CSP headers
+            // script-src 'self'
+            //      'sha384-Zao7ExRXVZOJobzS/uMp0P1jtJz3TTqJU4nYXkdmsjpiVD+/wcwCyX7FGqRIqvIz'
+            //      'sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM';
 
-    let js_directives = all_js_files
-        .into_iter()
-        .map(|f| f.hash)
-        .collect::<Vec<String>>();
+            let js_directives = all_js_files
+                .into_iter()
+                .map(|f| f.hash)
+                .collect::<Vec<String>>();
 
-    let js_checksums: String = js_directives
-        .iter()
-        .fold(String::new(), |mut output, value| {
-            let _ = write!(output, " 'sha384-{value}'");
-            output
-        });
+            let js_checksums: String =
+                js_directives
+                    .iter()
+                    .fold(String::new(), |mut output, value| {
+                        let _ = write!(output, " 'sha384-{value}'");
+                        output
+                    });
+
+            js_checksums
+        }
+    };
 
     let csp_header = format!(
         concat!(
