+++ server/testkit/tests/testkit/http_manifest.rs
@@ -1,6 +1,7 @@
 use kanidm_client::{http::header, KanidmClient};
+use kanidmd_core::config::ServerRole;
 
-#[kanidmd_testkit::test]
+#[kanidmd_testkit::test(role = ServerRole::WriteReplica)]
 async fn test_https_manifest(rsclient: &KanidmClient) {
     // We need to do manual reqwests here.
     let client = rsclient.client();
